refactor: move encode_instructions to own .c file
